Refactor entity IO, storage and handling (#548)
Part of #507 and part of #468.

+++ TREnvironmentEditor/Model/Types/Entities/EMMoveSecretFunction.cs
@@ -16,7 +16,7 @@ public class EMMoveSecretFunction : EMMovePickupFunction
             int entityIndex = location.GetContainedSecretEntity(level, floorData);
             if (entityIndex != -1)
             {
-                Types.Add(level.Entities[entityIndex].TypeID);
+                Types.Add((short)level.Entities[entityIndex].TypeID);
             }
         }
 
@@ -33,7 +33,7 @@ public class EMMoveSecretFunction : EMMovePickupFunction
             int entityIndex = location.GetContainedSecretEntity(level, floorData);
             if (entityIndex != -1)
             {
-                Types.Add(level.Entities[entityIndex].TypeID);
+                Types.Add((short)level.Entities[entityIndex].TypeID);
             }
         }
 
@@ -50,7 +50,7 @@ public class EMMoveSecretFunction : EMMovePickupFunction
             int entityIndex = location.GetContainedSecretEntity(level, floorData);
             if (entityIndex != -1)
             {
-                Types.Add(level.Entities[entityIndex].TypeID);
+                Types.Add((short)level.Entities[entityIndex].TypeID);
             }
         }
